requirements/base.txt
# Master branches
https://github.com/django/django/archive/main.tar.gz
https://github.com/encode/django-rest-framework/archive/master.tar.gz

whitenoise==5.2.0
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
mechanicalsoup
django-environ
black
celery[redis]==4.4.6
django-cryptography
mailjet_rest
isort
flake8